LayoutTests/imported/w3c/web-platform-tests/IndexedDB/nested-cloning-small-expected.txt


PASS small typed array 
PASS blob 
PASS blob with small typed array 
FAIL blob with small typed array with key generator assert_equals: IndexedDB result class should match put() argument expected "[object Blob]" but got "[object Null]"
PASS blob array 
FAIL blob array with key generator assert_equals: IndexedDB result class should match put() argument expected "[object Blob]" but got "[object Null]"
PASS array of blobs and small typed arrays 
FAIL array of blobs and small typed arrays with key generator assert_equals: IndexedDB result class should match put() argument expected "[object Blob]" but got "[object Null]"
